requirements.txt
streamlit==1.29.0
python-dotenv==1.0.0
httpx==0.25.2
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0